Implementada regras de negocio pedidas no notion, implementadas rotas do endpoint /actions/transactions METHOD post e get

+++ app/back-end/src/database/migrations/20221115152512-create-account.js
@@ -10,7 +10,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       balance: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(12, 2),
         defaultValue: 100.00
       }
   });
